# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mona",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/interviews/{interview_id}": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "Get Interview",
        "operationId": "getInterview",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "List Interviews",
        "operationId": "listInterviews",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "generating",
                    "ready",
                    "done",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "seed",
                    "debrief",
                    "on_demand"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Interview_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "interviews"
        ],
        "summary": "Create Interview",
        "operationId": "createInterview",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewCreated"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/interviews/{interview_id}/questions/{question_id}/answer": {
      "post": {
        "tags": [
          "interviews"
        ],
        "summary": "Answer Question",
        "operationId": "answerQuestion",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResult"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/interviews/{interview_id}/questions/{question_id}/skip": {
      "post": {
        "tags": [
          "interviews"
        ],
        "summary": "Skip Question",
        "operationId": "skipQuestion",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewQuestion"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/{interview_id}/questions/{question_id}/apply": {
      "post": {
        "tags": [
          "interviews"
        ],
        "summary": "Apply Question",
        "operationId": "applyQuestionRules",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyAllResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/{interview_id}/cancel": {
      "post": {
        "tags": [
          "interviews"
        ],
        "summary": "Cancel Interview",
        "operationId": "cancelInterview",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionBody": {
        "properties": {
          "conversationId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^cnv_[0-9a-hjkmnp-tv-z]{26}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversationid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ActionBody"
      },
      "AnswerBody": {
        "properties": {
          "optionId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Optionid"
          },
          "freeText": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Freetext"
          },
          "conversationId": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^cnv_[0-9a-hjkmnp-tv-z]{26}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversationid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AnswerBody"
      },
      "AnswerResult": {
        "additionalProperties": true,
        "type": "object"
      },
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApplyAllResult": {
        "additionalProperties": true,
        "type": "object"
      },
...
      "BatchScope-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "batch",
            "title": "Type"
          },
          "batchId": {
            "type": "string",
            "title": "Batchid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "batchId"
        ],
        "title": "BatchScope"
      },
...
      "CounterpartyScope-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "counterparty",
            "title": "Type"
          },
          "counterpartyId": {
            "type": "string",
            "title": "Counterpartyid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "counterpartyId"
        ],
        "title": "CounterpartyScope"
      },
...
      "DocumentsScope-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "documents",
            "title": "Type"
          },
          "documentIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Documentids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "documentIds"
        ],
        "title": "DocumentsScope"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Interview": {
        "additionalProperties": true,
        "type": "object"
      },
...
      "InterviewCreate": {
        "properties": {
          "scope": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/BatchScope-Input"
              },
              {
                "$ref": "#/components/schemas/QueueScope-Input"
              },
              {
                "$ref": "#/components/schemas/CounterpartyScope-Input"
              },
              {
                "$ref": "#/components/schemas/DocumentsScope-Input"
              },
              {
                "$ref": "#/components/schemas/SeedScope-Input"
              }
            ],
            "title": "Scope",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "batch": "#/components/schemas/BatchScope-Input",
                "counterparty": "#/components/schemas/CounterpartyScope-Input",
                "documents": "#/components/schemas/DocumentsScope-Input",
                "queue": "#/components/schemas/QueueScope-Input",
                "seed": "#/components/schemas/SeedScope-Input"
              }
            }
          },
          "lang": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "en",
                  "fr",
                  "ro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scope"
        ],
        "title": "InterviewCreate"
      },
      "InterviewCreated": {
        "additionalProperties": true,
        "type": "object"
      },
...
      "InterviewQuestion": {
        "additionalProperties": true,
        "type": "object"
      },
...
      "Page_Interview_": {
        "additionalProperties": true,
        "type": "object"
      },
      "QueueScope-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "queue",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "QueueScope"
      },
...
      "SeedScope-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "seed",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "SeedScope"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}